% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{data}
\name{aCharts}
\alias{aCharts}
\title{acharts R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
aCharts
}
\description{
Create virtual reality charts.
}
\section{Methods}{

\itemize{
  \item{\code{scatter} Scatter plot.}
  \item{\code{bar} Bar chart.}
  \item{\code{line} Line chart.}
  \item{\code{labels} Add labels.}
  \item{\code{plot} Plot chart}
  \item{\code{get} Get chart}
  \item{\code{browse} Browse chart}
  \item{\code{embed} Embed chart}
  \item{\code{insert} Insert \code{aframer} components in the \code{a_scene}.}
}
}

\section{Arguments}{


\itemize{
  \item{\code{scatter}, required: \code{data}, \code{x}, \code{y}, \code{z}, optional: \code{size}, \code{color}}
  \item{\code{bar}, required: \code{data}, \code{x}, \code{y}, \code{z}, \code{height}, optional: \code{color}, \code{width}}
  \item{\code{line}, required: \code{data}, \code{src.x}, \code{src.y}, \code{src.z}, \code{tgt.x}, \code{tgt.y}, \code{tgt.z}, optional: \code{color}, \code{width}}
  \item{\code{labels}, required: \code{data}, \code{x}, \code{y}, \code{z}, \code{text}, optional: \code{color}}
}
}

\section{Details}{


\code{insert} is used to "insert" aframe tags in the scene (\link[aframer]{a_scene})
}

\examples{
aCharts$
  new()$
  scatter(
    data = mtcars,
    x = mpg,
    y = drat,
    z = wt,
    color = qsec
  )$
  plot()$
  browse()

}
\keyword{datasets}
