% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{data}
\name{aCharts}
\alias{aCharts}
\title{acharts R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
aCharts
}
\description{
Create virtual reality charts.
}
\section{Methods}{

\itemize{
  \item{\code{scatter} Scatter plot.}
  \item{\code{bar} Bar chart.}
  \item{\code{line} Line chart.}
  \item{\code{labels} Add labels.}
  \item{\code{plot} Plot chart}
  \item{\code{get} Get chart}
  \item{\code{browse} Browse chart}
  \item{\code{embed} Embed chart}
  \item{\code{insert} Insert \code{aframer} components in the \code{a_scene}.}
}
}

\section{Arguments}{

The core functions (\code{scatter}, \code{bar} and \code{labels}) all take,
\code{data}, \code{x}, \code{y} and \code{z} as first and required arguments.
\code{bar} also requires \code{height} and \code{labels} requires \code{text}.
A \code{colors} palette can be passed when initialising the class (\code{new}).
}

\examples{
aCharts$
  new()$
  scatter(
    data = mtcars,
    x = mpg,
    y = drat,
    z = wt,
    color = qsec
  )$
  plot()$
  browse()

}
\keyword{datasets}
